fixed overflow distance on sensorState with turning value3 of mqttmsg into an int, added sensor state test

+++ Partial Component Demo/Pixy2/mqtt_queue.h
@@ -17,7 +17,7 @@ typedef struct
     uint8_t type;
     uint8_t value1;
     uint8_t value2;
-    uint8_t value3;
+    int value3;
     int value4;
 } MQTTMsg;
